refactor(parse-multi): split py.rs helpers into py_extract
Audit follow-up (2026-05- finding:               src/py.rs:49 Refactor LOW 12
py.rs is 291 lines and close to the 300-line cap).

Extracts the private helpers (extract_methods, extract_docstring,
unwrap_decorated, py_kind_to_item_kind, py_extract_name, should_skip)
into a sibling py_extract module. py.rs shrinks from 299 to 177 lines;
the new module is 179 lines. Each is well under the 300-line Rust cap
and has room to grow.

Public API is preserved: py::should_skip and py::extract_docstring
remain at their existing paths via pub use re-exports, so the
integration test (tests/py_parser.rs) needs no change.

Tests:
 30 passed (15 unit incl. 2 new
  in py_extract::tests + 15 integration)
- cargo clippy -p convergio-parse-multi --all-targets -- -D warnings
