+++ crates/convergio-parse-multi/src/py.rs
@@ -20,13 +20,21 @@
 //!
 //! If the first non-comment statement in a function or class body is a
 //! bare string literal (`expression_statement` → `string`), it is
-//! captured via [`extract_docstring`]. Currently used for tracing; callers
-//! may invoke it directly for indexing.
+//! captured via [`crate::py_extract::extract_docstring`]. Currently used
+//! for tracing; callers may invoke it directly for indexing.
 //!
 //! # Skip patterns
 //!
 //! Files whose path contains `__pycache__` or `.venv` return
-//! `Ok((vec![], vec![]))` immediately so the fleet walker can filter cheaply.
+//! `Ok((vec![], vec![]))` immediately so the fleet walker can filter
+//! cheaply via [`crate::py_extract::should_skip`].
+//!
+//! # Module split
+//!
+//! Helpers (`extract_methods`, `extract_docstring`, `unwrap_decorated`,
+//! `py_kind_to_item_kind`, `py_extract_name`, `should_skip`) live in
+//! `py_extract.rs` to keep this file under the 300-line cap; see the
+//! 2026-05-12 audit follow-up.
 
 use convergio_graph::model::{Edge, EdgeKind, Node, NodeKind};
 use tree_sitter::Parser;
@@ -34,7 +42,11 @@ use tree_sitter::Parser;
 use crate::{
     error::{ParseError, Result},
     lang::Lang,
+    py_extract::{extract_methods, py_extract_name, py_kind_to_item_kind, unwrap_decorated},
 };
+
+#[doc(inline)]
+pub use crate::py_extract::{extract_docstring, should_skip};
 
 /// Parse a Python source file and return graph-compatible nodes and edges.
 ///
@@ -162,138 +174,4 @@ pub fn parse_py(repo_name: &str, file_path: &str, source: &[u8]) -> Result<(Vec<
     }
 
     Ok((nodes, edges))
-}
-
-fn extract_methods(
-    class_node: tree_sitter::Node<'_>,
-    source: &str,
-    repo_name: &str,
-    file_path: &str,
-    class_id: &str,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-) {
-    let Some(body) = class_node.child_by_field_name("body") else {
-        return;
-    };
-    let mut cursor = body.walk();
-    for child in body.children(&mut cursor) {
-        let decl = unwrap_decorated(child);
-        if decl.kind() != "function_definition" {
-            continue;
-        }
-        let Some(name) = py_extract_name(decl, source) else {
-            continue;
-        };
-
-        let start = decl.start_byte() as u32;
-        let end = decl.end_byte() as u32;
-        let span = Some((start, end));
-        let node_id = Node::compute_id(
-            NodeKind::Item,
-            repo_name,
-            repo_name,
-            Some(file_path),
-            &name,
-            span,
-        );
-
-        tracing::debug!(
-            file = file_path,
-            item_kind = "method",
-            name = %name,
-            docstring = extract_docstring(decl, source).as_deref().unwrap_or(""),
-            "extracted py method"
-        );
-
-        edges.push(Edge {
-            src: class_id.to_owned(),
-            dst: node_id.clone(),
-            kind: EdgeKind::Declares,
-            weight: 1,
-        });
-        nodes.push(Node {
-            id: node_id,
-            kind: NodeKind::Item,
-            name,
-            file_path: Some(file_path.to_owned()),
-            crate_name: repo_name.to_owned(),
-            repo: repo_name.to_owned(),
-            item_kind: Some("method"),
-            span,
-        });
-    }
-}
-
-/// Return `true` for paths that should not be parsed.
-pub fn should_skip(file_path: &str) -> bool {
-    file_path.contains("__pycache__") || file_path.contains(".venv")
-}
-
-/// Extract the docstring from a function or class body, if present.
-///
-/// Returns the raw string token (including quotes) of the first
-/// `expression_statement → string` in the body, skipping leading comments.
-pub fn extract_docstring(node: tree_sitter::Node<'_>, source: &str) -> Option<String> {
-    let body = node.child_by_field_name("body")?;
-    let mut cursor = body.walk();
-    for child in body.children(&mut cursor) {
-        match child.kind() {
-            "comment" => continue,
-            "expression_statement" => {
-                let mut expr_cursor = child.walk();
-                for expr in child.children(&mut expr_cursor) {
-                    if expr.kind() == "string" {
-                        return source
-                            .get(expr.start_byte()..expr.end_byte())
-                            .map(str::to_owned);
-                    }
-                }
-                return None;
-            }
-            _ => return None,
-        }
-    }
-    None
-}
-
-fn unwrap_decorated(node: tree_sitter::Node<'_>) -> tree_sitter::Node<'_> {
-    if node.kind() != "decorated_definition" {
-        return node;
-    }
-    if let Some(def) = node.child_by_field_name("definition") {
-        return def;
-    }
-    let mut cursor = node.walk();
-    for child in node.children(&mut cursor) {
-        if py_kind_to_item_kind(child.kind()).is_some() {
-            return child;
-        }
-    }
-    node
-}
-
-fn py_kind_to_item_kind(kind: &str) -> Option<&'static str> {
-    match kind {
-        "function_definition" => Some("function"),
-        "class_definition" => Some("class"),
-        _ => None,
-    }
-}
-
-fn py_extract_name(node: tree_sitter::Node<'_>, source: &str) -> Option<String> {
-    if let Some(name_node) = node.child_by_field_name("name") {
-        return source
-            .get(name_node.start_byte()..name_node.end_byte())
-            .map(str::to_owned);
-    }
-    let mut cursor = node.walk();
-    for child in node.children(&mut cursor) {
-        if child.kind() == "identifier" {
-            return source
-                .get(child.start_byte()..child.end_byte())
-                .map(str::to_owned);
-        }
-    }
-    None
 }
